fp synth: keep glottal source float to kill low-VoicingGain quantization noise

+++ include/KlattSynthesizerFP.h
@@ -143,7 +143,7 @@ namespace SharpVox {
         int32_t _f5pD1, _f5pD2;
         int32_t _f6pD1, _f6pD2;
         int32_t _preemphPrev;   // pre-emphasis one-sample delay, Q0
-        int32_t _tiltPrev;      // spectral tilt one-sample delay, Q0
+        float _tiltPrev;        // spectral tilt one-sample delay (float, matches float variant)
 
         //  Amplitude state: Q8 (float * 256) 
         int32_t _voiceAmp_q8;
